MPAE-18930 ported project to use latest MCC Melody and XC8 v3.0. Added functionality of using CLCs and Timer for switch debouncing. Added functionality to EEPROM lab to only write to EEPROM after switch 2 is pressed.

+++ pic18f47q84-curiosity-hpc-demo-code.X/Labs/Lab03_Rotate/Rotate.c
@@ -36,7 +36,7 @@
     TERMS.
 */
 
-#include "../../mcc_generated_files/pin_manager.h"
+#include "../../mcc_generated_files/system/system.h"
 #include "../../labs.h"
 
 static uint8_t rotateReg;
@@ -61,7 +61,7 @@ void Rotate(void) {
             rotateReg = 1;
 
         
-        LEDs = (rotateReg << 4);                                                // Determine which LED will light up
+        LEDs = (uint8_t)(rotateReg << 4);                                                // Determine which LED will light up
     }                                                                           // ie. which bit in the register the 1 has rotated to.
 
     if (switchEvent) {
